Revert "chore: break out CAN utils into new header file"
This reverts commit 44ca6414cb27a2f9df24f42de654476bf6ce19cb.
