--- common/meson.build
@@ -1,5 +1,0 @@
-common_inc = include_directories('include')
-common_src = files(
-  'src' / 'can-utils.c',
-  'src' / 'utils.c',
-)
