+++ servo-node/Core/Src/main.c
@@ -31,14 +31,9 @@
 #include <string.h>
 
 #include "app.h"
-
-// FreeRTOS includes
 #include "queue.h"
 #include "task.h"
 #include "timers.h"
-
-// Common utils
-#include "can-utils.h"
 #include "utils.h"
 
 /* USER CODE END Includes */
@@ -126,6 +121,7 @@ void StartDefaultTask(void *argument);
 
 /* USER CODE BEGIN PFP */
 
+void StartCANTxTask(void *argument);
 void StartMeasureTask(void *argument);
 
 /* USER CODE END PFP */
@@ -207,8 +203,7 @@ int main(void) {
 
   /* USER CODE BEGIN RTOS_THREADS */
 
-  CANTxTaskHandle =
-      osThreadNew(StartCANTxTask, &CANMessageQueue, &CANTxTask_attributes);
+  CANTxTaskHandle = osThreadNew(StartCANTxTask, NULL, &CANTxTask_attributes);
   measureTaskHandle =
       osThreadNew(StartMeasureTask, NULL, &measureTask_attributes);
 
@@ -832,6 +827,52 @@ void defaultTaskTimer(TimerHandle_t xTimer) {
 void measureTaskTimer(TimerHandle_t xTimer) {
   UNUSED(xTimer);
   SignalTask(measureTaskHandle);
+}
+
+void mailboxFreeCallback(CAN_HandleTypeDef *_hcan) {
+  // Only signal when going from 0 free mailboxes to 1 free
+  if (HAL_CAN_GetTxMailboxesFreeLevel(_hcan) <= 1) {
+    SignalTask(CANTxTaskHandle);
+  }
+}
+
+void HAL_CAN_TxMailbox0CompleteCallback(CAN_HandleTypeDef *_hcan) {
+  mailboxFreeCallback(_hcan);
+}
+
+void HAL_CAN_TxMailbox1CompleteCallback(CAN_HandleTypeDef *_hcan) {
+  mailboxFreeCallback(_hcan);
+}
+
+void HAL_CAN_TxMailbox2CompleteCallback(CAN_HandleTypeDef *_hcan) {
+  mailboxFreeCallback(_hcan);
+}
+
+void StartCANTxTask(void *argument) {
+  UNUSED(argument);
+
+  HAL_CAN_ActivateNotification(&hcan, CAN_IT_TX_MAILBOX_EMPTY);
+  HAL_CAN_Start(&hcan);
+
+  uint32_t mailbox = 0;
+
+  for (;;) {
+    if (HAL_CAN_GetTxMailboxesFreeLevel(&hcan) == 0) {
+      // Wait for mailbox to become available
+      ulTaskNotifyTake(pdTRUE, portMAX_DELAY);
+    }
+
+    CANFrame frame;
+    // Wait for queue to receive message
+    xQueueReceive(CANMessageQueue, &frame, portMAX_DELAY);
+
+    CAN_TxHeaderTypeDef header = {
+        .StdId = frame.id,
+        .DLC = frame.dlc,
+    };
+
+    HAL_CAN_AddTxMessage(&hcan, &header, frame.data, &mailbox);
+  }
 }
 
 void StartMeasureTask(void *argument) {
